+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/typed/CanBeSubclassDeterminer.scala
@@ -122,6 +122,10 @@ trait CanBeSubclassDeterminer {
   }
 
   private def canBeSubclassOf(givenTypes: Set[SingleTypingResult], superclassCandidates: Set[SingleTypingResult]): ValidatedNel[String, Unit] = {
+    // Would be more safety to do givenTypes.forAll(... superclassCandidates.exists ...) - we wil protect against
+    // e.g. (String | Int).canBeSubclassOf(String) which can fail in runtime for Int, but on the other hand we can't block user's intended action.
+    // He/she could be sure that in this type, only String will appear. He/she also can't easily downcast (String | Int) to String so leaving here
+    // "double exists" looks like a good tradeoff
     condNel(givenTypes.exists(given => superclassCandidates.exists(singleCanBeSubclassOf(given, _).isValid)), (),
       s"""None of the following types:
          |${givenTypes.map(" - " + _.display).mkString(",\n")}